[DayPicker] Rename DatePickerView to DayPickerView and move to own module
