+++ packages/material-ui-lab/src/ClockPicker/index.ts
@@ -3,3 +3,4 @@ export { default } from './ClockPickerStandalone';
 export type ClockPickerProps<
   TDate
 > = import('./ClockPickerStandalone').ClockPickerStandaloneProps<TDate>;
+export type ClockPickerView = NonNullable<ClockPickerProps<unknown>['view']>;
