+++ packages/material-ui-lab/src/DayPicker/shared.ts
@@ -1,0 +1,1 @@
+export type DayPickerView = 'year' | 'date' | 'month';
