+++ packages/material-ui-lab/src/internal/pickers/typings/Views.ts
@@ -1,3 +1,1 @@
 export type AllAvailableViews = 'year' | 'date' | 'month' | 'hours' | 'minutes' | 'seconds';
-
-export type DatePickerView = 'year' | 'date' | 'month';
